requirements.txt
alpaca-trade-api==0.53.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Authlib==0.15.3
babel==2.9.0
backcall==0.2.0
backtrader==1.9.76.123
bleach==3.3.0
Brotli==1.0.9
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cssselect2==0.4.1
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastapi==0.63.0
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
h11==0.12.0
html5lib==1.1
httpcore==0.12.3
httpx==0.17.0
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lab==6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.8.0
pyparsing==2.4.7
Pyphen==0.10.0
pyreadline==2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
rfc3986==1.4.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
tda-api==1.1.7
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tinycss2==1.1.0
tornado==6.1
traitlets==5.0.5
tulipy==0.4.0
txt2tags==3.7
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
wcwidth==0.2.5
WeasyPrint==0.42.3
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0